+++ certbot-compatibility-test/certbot_compatibility_test/configurators/apache/common.py
@@ -2,11 +2,8 @@
 import os
 import shutil
 import subprocess
+from unittest import mock
 
-try:
-    import mock
-except ImportError: # pragma: no cover
-    from unittest import mock # type: ignore
 import zope.interface
 
 from certbot import errors as le_errors