+++ certbot-nginx/certbot_nginx/_internal/parser_obj.py
@@ -4,7 +4,6 @@ raw lists of tokens from pyparsing. """
 import abc
 import logging
 
-import six
 
 from acme.magic_typing import List
 from certbot import errors
@@ -152,7 +151,7 @@ class Statements(Parsable):
         if not isinstance(raw_list, list):
             raise errors.MisconfigurationError("Statements parsing expects a list!")
         # If there's a trailing whitespace in the list of statements, keep track of it.
-        if raw_list and isinstance(raw_list[-1], six.string_types) and raw_list[-1].isspace():
+        if raw_list and isinstance(raw_list[-1], str) and raw_list[-1].isspace():
             self._trailing_whitespace = raw_list[-1]
             raw_list = raw_list[:-1]
         self._data = [parse_raw(elem, self, add_spaces) for elem in raw_list]
@@ -184,7 +183,7 @@ class Statements(Parsable):
 def _space_list(list_):
     """ Inserts whitespace between adjacent non-whitespace tokens. """
     spaced_statement = [] # type: List[str]
-    for i in reversed(six.moves.xrange(len(list_))):
+    for i in reversed(range(len(list_))):
         spaced_statement.insert(0, list_[i])
         if i > 0 and not list_[i].isspace() and not list_[i-1].isspace():
             spaced_statement.insert(0, " ")
@@ -206,7 +205,7 @@ class Sentence(Parsable):
         :returns: whether this lists is parseable by `Sentence`.
         """
         return isinstance(lists, list) and len(lists) > 0 and \
-            all(isinstance(elem, six.string_types) for elem in lists)
+            all(isinstance(elem, str) for elem in lists)
 
     def parse(self, raw_list, add_spaces=False):
         """ Parses a list of string types into this object.
@@ -214,7 +213,7 @@ class Sentence(Parsable):
         if add_spaces:
             raw_list = _space_list(raw_list)
         if not isinstance(raw_list, list) or \
-                any(not isinstance(elem, six.string_types) for elem in raw_list):
+                any(not isinstance(elem, str) for elem in raw_list):
             raise errors.MisconfigurationError("Sentence parsing expects a list of string types.")
         self._data = raw_list