+++ tests/letstest/scripts/test_leauto_upgrades.sh
@@ -33,19 +33,22 @@ if ! ./letsencrypt-auto -v --debug --version --no-self-upgrade 2>&1 | tail -n1 |
     exit 1
 fi
 
-# This script sets the environment variables PYTHON_NAME, VENV_PATH, and
-# VENV_SCRIPT based on the version of Python available on the system. For
-# instance, Fedora uses Python 3 and Python 2 is not installed.
-. tests/letstest/scripts/set_python_envvars.sh
+if command -v python; then
+    PYTHON_NAME="python"
+else
+    PYTHON_NAME="python3"
+fi
 
 # Now that python and openssl have been installed, we can set up a fake server
 # to provide a new version of letsencrypt-auto. First, we start the server and
 # directory to be served.
 MY_TEMP_DIR=$(mktemp -d)
 PORT_FILE="$MY_TEMP_DIR/port"
+LOG_FILE="$MY_TEMP_DIR/log"
 SERVER_PATH=$("$PYTHON_NAME" tools/readlink.py tools/simple_http_server.py)
 cd "$MY_TEMP_DIR"
-"$PYTHON_NAME" "$SERVER_PATH" 0 > $PORT_FILE &
+# We set PYTHONUNBUFFERED to disable buffering of output to LOG_FILE
+PYTHONUNBUFFERED=1 "$PYTHON_NAME" "$SERVER_PATH" 0 > $PORT_FILE 2> "$LOG_FILE" &
 SERVER_PID=$!
 trap 'kill "$SERVER_PID" && rm -rf "$MY_TEMP_DIR"' EXIT
 cd ~-
@@ -118,4 +121,49 @@ fi
 if ! diff letsencrypt-auto letsencrypt-auto-source/letsencrypt-auto ; then
     echo letsencrypt-auto and letsencrypt-auto-source/letsencrypt-auto differ
     exit 1
+fi
+
+# Now let's test if letsencrypt-auto still tries to upgrade to a new version.
+# Regardless of the OS, versions of the script with development version numbers
+# ending in .dev0 will not upgrade. See
+# https://github.com/certbot/certbot/blob/bdfb9f19c4086a60ef010d2431768850c26d838a/certbot-auto#L1947-L1948.
+# In order to test the process of different OSes setting NO_SELF_UPGRADE as
+# part of the script's deprecation, we make use of the fact that
+# letsencrypt-auto should still attempt to fetch the version number from PyPI
+# even if it has a development version number unless NO_SELF_UPGRADE is set in
+# which case all of that logic should be skipped.
+#
+# First we make a copy of the current server logs.
+PREVIOUS_LOG_FILE="$MY_TEMP_DIR/previous-log"
+cp "$LOG_FILE" "$PREVIOUS_LOG_FILE"
+
+# Next we run letsencrypt-auto and make sure there were no problems checking
+# for updates, the Certbot install still works, the version number is what
+# we expect, and it prints a message about not receiving updates.
+if ./letsencrypt-auto -v --debug --version | grep "WARNING: couldn't find Python" ; then
+    echo "Had problems checking for updates!"
+    exit 1
+fi
+if ! ./letsencrypt-auto -v --debug --version 2>&1 | tail -n1 | grep "^certbot $EXPECTED_VERSION$" ; then
+    echo unexpected certbot version found
+    exit 1
+fi
+if ! ./letsencrypt-auto -v --debug --version 2>&1 | grep "will no longer receive updates" ; then
+    echo script did not print warning about not receiving updates!
+    exit 1
+fi
+
+# Finally, we check if our local server received more requests. Over time,
+# we'll move more and more OSes into this case until it this is the expected
+# behavior on all systems.
+if [ -f /etc/redhat-release ]; then
+    if ! diff "$LOG_FILE" "$PREVIOUS_LOG_FILE" ; then
+        echo our local server received unexpected requests
+        exit 1
+    fi
+else
+    if diff "$LOG_FILE" "$PREVIOUS_LOG_FILE" ; then
+        echo our local server did not receive the requests we expected
+        exit 1
+    fi
 fi